pyproject.toml
[project]
name = "consult-app-api"
version = "0.1.0"
description = "Consult App CI/CD"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "black>=24.1.1",
    "fastapi>=0.111.0",
    "flake8>=7.0.0",
    "httpx>=0.26.0",
    "pre-commit>=3.6.0",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest>=7.4.4",
    "python-multipart>=0.0.7",
    "requests>=2.31.0",
    "uvicorn>=0.27.0",
    "google-genai>=1.39.1",
    "google-cloud-storage>=2.18.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "openai>=1.95.1",
    "openai-agents>=0.3.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.1",
    "chromadb>=0.5.20",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["llm_api*"]
